Move PlaceAutocompleteView to internal package
