+++ mapzen-places-api/src/main/res/layout/place_autocomplete_fragment.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.mapzen.places.api.ui.PlaceAutocompleteView
+<com.mapzen.places.api.internal.PlaceAutocompleteView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"/>
